requirements.txt
aniso8601==6.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
Click==7.0
et-xmlfile==1.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Excel==0.0.7
Flask-JWT-Extended==3.18.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==20.1.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.10.1
lml==0.1.0
MarkupSafe==1.1.1
openpyxl==3.0.9
pyaml==21.10.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pymodbus==2.5.3
pyserial==3.5
pytz==2019.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.3.3
texttable==1.6.4
urllib3==1.26.8
Werkzeug==1.0.1
XlsxWriter==3.0.3
